adding endpoint for all user fetching
